# public/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CGS Games API",
    "version": "1.0.0",
    "description": "Discover community Card Game Simulator games. Public reads require no authentication. Publishing and deletion use Firebase ID tokens and a user profile with a username. API errors retain the legacy error string and add code, message, and hint. GET routes also support HEAD; OPTIONS lists allowed methods. Unknown API paths return JSON 404; unsupported methods return JSON 405 with Allow."
  },
  "servers": [
    {
      "url": "https://cgs.games"
    }
  ],
...
  "security": [],
  "paths": {
...
    "/api/games/upload": {
      "post": {
        "operationId": "uploadGame",
        "summary": "Publish a .cgs.zip archive (maximum 100 MiB)",
        "description": "Send a multipart file directly, or first upload the archive to Firebase Storage using an authenticated client and then submit its staged path. A staged archive is cleaned up after processing is attempted. Archives need a valid cgs.json and supported game assets.",
        "security": [
          {
            "firebaseIdToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": ".cgs.zip archive, at most 104857600 bytes."
                  }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["stagedPath", "originalFilename"],
                "properties": {
                  "stagedPath": {
                    "type": "string",
                    "pattern": "^staged-uploads/[^/]+/[A-Za-z0-9_-]+\\.cgs\\.zip$",
                    "description": "staged-uploads/{authenticated Firebase uid}/{unique-id}.cgs.zip"
                  },
                  "originalFilename": {
                    "type": "string",
                    "pattern": "\\.cgs\\.zip$"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Game published successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishedGame"
                }
              }
            }
          },
          "400": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Allow": {
                "description": "Allowed HTTP methods.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Structured error. See code, message, and hint for recovery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "firebaseIdToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase Auth ID token from the signed-in user. Requires a creator username."
      },
      "warmToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Operator GATCG_WARM_TOKEN secret. Only for spoiler cache maintenance; never a Firebase ID token."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error", "code", "message", "hint"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Legacy error message retained for existing clients."
          },
          "code": {
            "type": "string",
            "example": "NOT_FOUND"
          },
          "message": {
            "type": "string"
          },
          "hint": {
            "type": "string"
          }
        }
      },
      "PublishedGame": {
        "type": "object",
        "required": ["success", "slug"],
        "properties": {
          "success": {
            "const": true
          },
          "slug": {
            "type": "string"
          }
        }
      },
...
    }
  }
}